refactor(backend): Remove dto folder, replace with .dto.ts file v2
