+++ apps/docs/content/docs/dev/database-filters.mdx
@@ -16,7 +16,7 @@ import { Step, Steps } from 'fumadocs-ui/components/steps';
  
 ### Arguments (DTO)
 
-```ts title="show/dto/show.args.ts"
+```ts title="show/show.dto.ts"
 @ArgsType()
 export class ShowExampleArgs extends PaginationArgs {
   @Field(() => String, { nullable: true }) // [!code highlight]
