+++ apps/docs/content/docs/dev/database-pagination.mdx
@@ -17,7 +17,7 @@ import { Step, Steps } from 'fumadocs-ui/components/steps';
 
 ### Object (DTO)
 
-```ts title="show/dto/show.args.ts"
+```ts title="show/show.dto.ts"
 import { Field, Int, ObjectType } from '@nestjs/graphql';
 import { PageInfo } from 'vitnode-backend'; // [!code highlight]
 
@@ -44,7 +44,7 @@ export class ShowTestObj {
 <Step>
 ### Arguments (DTO)
 
-```ts title="show/dto/show.args.ts"
+```ts title="show/show.dto.ts"
 import { ArgsType } from '@nestjs/graphql';
 import { PaginationArgs } from 'vitnode-backend'; // [!code highlight]
