+++ apps/docs/content/docs/dev/dto.mdx
@@ -14,7 +14,7 @@ import { Step, Steps } from 'fumadocs-ui/components/steps';
 
 You can pass arguments to the service method by creating a DTO (Data Transfer Object) file.
 
-```ts title="backend/plugins/blog/categories/show/dto/show.args.ts"
+```ts title="backend/plugins/blog/categories/show/show.dto.ts"
 import { ArgsType, Field } from '@nestjs/graphql';
 
 @ArgsType()
