+++ apps/docs/content/docs/dev/i18n.mdx
@@ -122,7 +122,7 @@ export const blog_categories_name = pgTable(
 
 Inside input DTO you can use `TextLanguageInput[]` for translation fields.
 
-```ts title="backend/plugins/blog/categories/show/dto/show.args.ts"
+```ts title="backend/plugins/blog/categories/show/show.dto.ts"
 import { ArgsType, Field } from '@nestjs/graphql';
 import { IsArray } from 'class-validator';
 import { Transform } from 'class-transformer';
