+++ packages/backend/src/core/admin/members/create/create.resolver.ts
@@ -1,4 +1,4 @@
-import { SignUpCoreSessionsObj } from '@/core/sessions/sign_up/dto/sign_up.obj';
+import { SignUpCoreSessionsObj } from '@/core/sessions/sign_up/sign_up.dto';
 import { AdminAuthGuards, GqlContext } from '@/utils';
 import { UseGuards } from '@nestjs/common';
 import { Args, Context, Mutation, Resolver } from '@nestjs/graphql';
