+++ packages/backend/src/core/admin/members/create/create.service.ts
@@ -1,5 +1,5 @@
-import { SignUpCoreSessionsObj } from '@/core/sessions/sign_up/dto/sign_up.obj';
 import { SignUpHelperService } from '@/core/sessions/sign_up/helpers/sign-up-helper.service';
+import { SignUpCoreSessionsObj } from '@/core/sessions/sign_up/sign_up.dto';
 import { GqlContext } from '@/utils';
 import { Injectable } from '@nestjs/common';
