+++ packages/backend/src/core/admin/sessions/authorization/authorization.dto.ts
@@ -1,4 +1,4 @@
-import { AuthorizationCurrentUserObj } from '@/core/sessions/authorization/dto/authorization.obj';
+import { AuthorizationCurrentUserObj } from '@/core/sessions/authorization/authorization.dto';
 import { Field, ObjectType } from '@nestjs/graphql';
 
 @ObjectType()
