+++ packages/backend/src/core/admin/styles/nav/edit/edit.resolver.ts
@@ -1,4 +1,4 @@
-import { ShowCoreNav } from '@/core/nav/show/dto/show.obj';
+import { ShowCoreNav } from '@/core/nav/show/show.dto';
 import { AdminAuthGuards } from '@/utils';
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
