+++ packages/backend/src/core/editor/upload/upload.service.ts
@@ -1,4 +1,7 @@
-import { UploadCoreFilesObj } from '@/core/files/helpers/upload/upload.dto';
+import {
+  UploadCoreFilesArgs,
+  UploadCoreFilesObj,
+} from '@/core/files/helpers/upload/upload.dto';
 import { FilesService } from '@/core/files/helpers/upload/upload.service';
 import { generateRandomString } from '@/functions/generate-random-string';
 import { InternalDatabaseService } from '@/utils/database/internal_database.service';
@@ -9,7 +12,6 @@ import { core_files } from '../../../database/schema/files';
 import { User } from '../../../decorators';
 import { AccessDeniedError } from '../../../errors';
 import { AllowTypeFilesEnum, getConfigFile } from '../../../providers/config';
-import { UploadCoreFilesArgs } from '../../files/helpers/upload/dto/upload.args';
 import {
   acceptMimeTypeImage,
   acceptMimeTypeVideo,
