--- packages/backend/src/core/nav/show/dto/show.args.ts
@@ -1,5 +1,0 @@
-import { PaginationArgs } from '@/utils';
-import { ArgsType } from '@nestjs/graphql';
-
-@ArgsType()
-export class ShowCoreNavArgs extends PaginationArgs {}
