+++ packages/backend/src/core/nav/show/show.service.ts
@@ -5,8 +5,7 @@ import { and, count, eq } from 'drizzle-orm';
 import { core_nav } from '../../../database/schema/nav';
 import { inputPaginationCursor, outputPagination } from '../../../functions';
 import { SortDirectionEnum } from '../../../utils';
-import { ShowCoreNavArgs } from './dto/show.args';
-import { ShowCoreNavObj } from './dto/show.obj';
+import { ShowCoreNavArgs, ShowCoreNavObj } from './show.dto';
 
 @Injectable()
 export class ShowCoreNavService {
