+++ packages/backend/src/core/sessions/authorization/authorization.dto.ts
