+++ packages/backend/src/core/sessions/devices/show/show.dto.ts
