+++ packages/backend/src/core/settings/show/show.service.ts
@@ -6,7 +6,7 @@ import { join } from 'path';
 import { ABSOLUTE_PATHS_BACKEND, getConfigFile } from '../../..';
 import { core_languages } from '../../../database/schema/languages';
 import { ManifestWithLang } from '../settings.module';
-import { ShowSettingsObj } from './dto/show.obj';
+import { ShowSettingsObj } from './show.dto';
 
 @Injectable()
 export class ShowSettingsService {
