+++ packages/backend/src/core/theme_editor/show/show.service.ts
@@ -8,7 +8,7 @@ import {
   ColorsShowCoreThemeEditor,
   HslColor,
   ShowCoreThemeEditorObj,
-} from './dto/show.obj';
+} from './show.dto';
 
 @Injectable()
 export class ShowCoreThemeEditorService {
